# contracts/ai-internal.openapi.yaml
openapi: 3.1.0
...
info:
  title: RAMALS internal AI contract
  version: 1.0.0
  summary: Non-authoritative agent proposals for the RAMALS deterministic core.
  description: |
    Every response on this API is a **proposal**. Nothing returned here is authoritative learner
    state. Spring validates each proposal against schema, semantics, trust level and deterministic
    policy before any authoritative write, and the write happens in a new Spring transaction.

    Authentication is a Keycloak client-credentials token carrying the `ramals-ai` audience
    (M1-ADR-003). A learner token is never accepted, and is never forwarded by Spring.
servers:
  - url: http://ramals-ai:8000
    description: In-cluster service address
security:
  - workloadIdentity: []
...
paths:
...
  /internal/v1/tutor/respond:
    post:
      operationId: tutorRespond
      tags: [agents]
      summary: Personalized explanation or hint
      description: Learner-facing instructional content only. Never alters mastery or evidence.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequestEnvelope'
      responses:
        '200':
          description: A non-authoritative tutor proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIProposalEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableProposal' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /internal/v1/diagnostic/propose:
    post:
      operationId: diagnosticPropose
      tags: [agents]
      summary: Propose the next diagnostic probe
      description: >
        Suggests what to probe next. Deterministic curriculum and prerequisite policy decides
        whether the proposal is eligible.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequestEnvelope'
      responses:
        '200':
          description: A diagnostic proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIProposalEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableProposal' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /internal/v1/assessment/propose:
    post:
      operationId: assessmentPropose
      tags: [agents]
      summary: Propose a candidate assessment item or rubric
      description: >
        Generated content is UNVERIFIED and cannot gate progression until promoted through approved
        validation or review.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequestEnvelope'
      responses:
        '200':
          description: An UNVERIFIED candidate item proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIProposalEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableProposal' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /internal/v1/assessment/evaluate:
    post:
      operationId: assessmentEvaluate
      tags: [agents]
      summary: Formative evaluation assistance
      description: |
        **FORMATIVE_ONLY (M1-ADR-010).** May return feedback, classifications, rubric suggestions and
        reviewer assistance. Must never create `ledger.evidence`, determine an authoritative score,
        change attempt completion, or affect mastery, confidence, recommendation or progression.

        The response `trustLevel` is always `FORMATIVE_ONLY`; any other value is a contract
        violation and Spring rejects it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequestEnvelope'
      responses:
        '200':
          description: A formative-only evaluation proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIProposalEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableProposal' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /internal/v1/adaptation/propose:
    post:
      operationId: adaptationPropose
      tags: [agents]
      summary: Propose the next pedagogical action
      description: >
        Deterministic RecommendationPolicy remains authoritative. Where the agent and the policy
        disagree, the policy wins and the disagreement is recorded for evaluation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequestEnvelope'
      responses:
        '200':
          description: An adaptation proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIProposalEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableProposal' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /internal/v1/diagnostic-assessment/propose:
    post:
      operationId: diagnosticAssessmentPropose
      tags: [agents]
      summary: Propose a grounded reading of what a learner's evidence shows
      description: >
        M2-T08/T09. Distinct from /internal/v1/diagnostic/propose, which proposes the next probe.
        This operation classifies skills from evidence and must cite it, so it takes a
        GroundedContext rather than a minimized learning context. Spring builds that context from the
        authenticated learner and remains the authority: DiagnosticAssessmentProposalGate decides
        whether the returned proposal has any effect.

        A separate operation rather than a new optional field on AIRequestEnvelope. The context is
        mandatory here and meaningless on the MVP-1 operations, and an optional field carrying that
        distinction would leave every existing operation with an ambiguous "is it required for me?"
        semantic that no schema states.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiagnosticAssessmentRequest'
      responses:
        '200':
          description: A diagnostic assessment proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIProposalEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableProposal' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /internal/v1/assessment-evaluation/propose:
    post:
      operationId: assessmentEvaluationPropose
      tags: [agents]
      summary: Propose rubric-bound evaluation of an AI-eligible learner response
      description: >
        M2-T11. This operation accepts only FREE_TEXT, DESIGN, or REASONING responses. MCQ,
        TRUE_FALSE, NUMERIC, EXACT_MATCH, and executable responses have no representation on this
        boundary and remain on Spring's deterministic scoring path. The response is a
        NON_AUTHORITATIVE AssessmentEvaluationProposal; Spring's M2-T12 gate owns any score or
        evidence effect.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentEvaluationRequest'
      responses:
        '200':
          description: A non-authoritative, rubric-bound assessment evaluation proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIProposalEnvelope'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableProposal' }
        '503': { $ref: '#/components/responses/AiUnavailable' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
components:
  securitySchemes:
    workloadIdentity:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak client-credentials token with audience `ramals-ai` (M1-ADR-003). A learner token
        carries the `ramals-api` audience and is rejected here.
  responses:
    BadRequest:
      description: The request envelope is malformed or violates a documented limit
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Missing, expired or wrong-audience workload token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Authenticated, but not the RAMALS core workload
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    UnprocessableProposal:
      description: No valid proposal could be produced within the bounded repair budget
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    AiUnavailable:
      description: >
        The model route or agent runtime is unavailable. Deterministic learning is unaffected; the
        caller degrades the AI feature rather than failing the learner's action.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    DeadlineExceeded:
      description: >
        The caller's absolute deadline elapsed (M1-ADR-001). Fails closed: no partial proposal.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    ContractVersion:
      type: string
      pattern: '^\d+\.\d+$'
      examples: ['1.0']
      description: Envelope contract version. A breaking change ships a new major version.
...
    AgentType:
      type: string
      enum: [TUTOR, DIAGNOSTIC, ASSESSMENT, ADAPTATION]
...
    LearnerRef:
      type: object
      description: >
        Opaque learner reference. Never a database primary key, and never enough on its own to
        identify a person outside the platform.
      additionalProperties: false
      required: [learnerRef]
      properties:
        learnerRef:
          type: string
          minLength: 1
          maxLength: 128
        locale:
          type: string
          maxLength: 16
          examples: ['en-IN']
    LearningContext:
      type: object
      description: >
        Minimized, purpose-specific learner state. Spring sends only what the agent use case needs;
        no full history, no raw keys, no access token.
      additionalProperties: false
      required: [skillCode]
      properties:
        # 96 matches core.skill.stable_code. The boundary was narrower than the authoritative
        # schema, so a skill code of 65-96 characters was storable, legal, and rejected here.
        skillCode:
          type: string
          minLength: 1
          maxLength: 96
        masteryScore:
          $ref: '#/components/schemas/DecimalString'
        evidenceConfidence:
          $ref: '#/components/schemas/DecimalString'
        masteryStatus:
          type: string
          maxLength: 32
          examples: ['NEEDS_PRACTICE']
        # Prerequisites are skill stable codes, so they carry the same width as skillCode. Widening
        # only skillCode would have left the identical defect one field over.
        prerequisites:
          type: array
          maxItems: 32
          items:
            type: string
            maxLength: 96
    DomainContext:
      type: object
      description: >
        Which learning domain this request is about, and which published curriculum version defines
        it. Present so an agent never has to assume the domain: the platform is domain-neutral by
        construction, and the first domain must not become the permanent one by accident.

        Spring resolves every field from authoritative state. An agent must treat these as facts
        about the request, not as hints it may reinterpret.
      additionalProperties: false
      required: [domainCode, domainType]
      properties:
        # 64 matches core.learning_domain.code. Deliberately not 96: each identifier carries the
        # width of its own column, rather than the widest column in the schema.
        domainCode:
          type: string
          minLength: 1
          maxLength: 64
          examples: ['KAFKA']
        domainType:
          $ref: '#/components/schemas/DomainType'
        # 64 matches core.curriculum_version.version_code.
        curriculumVersion:
          type: string
          maxLength: 64
          examples: ['v1']
...
    LearningGoalContext:
      type: object
      description: >
        What the learner is working towards, when that is known and relevant to the request.

        Optional throughout, because the authoritative learner model carries a single goal today and
        richer goal structures are an MVP-5 concern. The shape exists now so that agents are written
        against a general goal rather than against "the learner's one domain", which is the
        assumption that would be expensive to unpick later.
      additionalProperties: false
      required: [goalType, goalCode]
      properties:
        goalType:
          $ref: '#/components/schemas/GoalType'
        # 96: a goal code may reference a skill-scale identifier.
        goalCode:
          type: string
          minLength: 1
          maxLength: 96
          examples: ['KAFKA']
        targetDate:
          type: string
          format: date
        goalVersion:
          type: string
          maxLength: 64
...
    AIRequestEnvelope:
      type: object
      additionalProperties: false
      required: [contractVersion, interactionId, requestId, learner, constraints]
      properties:
        contractVersion: { $ref: '#/components/schemas/ContractVersion' }
        interactionId:
          type: string
          maxLength: 64
          description: Logical learner action. Stable across safe retries.
        requestId:
          type: string
          maxLength: 64
          description: One transport attempt. New per retry.
        learner: { $ref: '#/components/schemas/LearnerRef' }
        learningContext: { $ref: '#/components/schemas/LearningContext' }
        # Optional rather than required: making a new property required would reject payloads a peer
        # already sends, which the compatibility gate correctly refuses. Spring populates
        # domainContext on every request it builds; the optionality is a contract-evolution
        # concession, not permission for a caller to omit it.
        domainContext: { $ref: '#/components/schemas/DomainContext' }
        learningGoalContext: { $ref: '#/components/schemas/LearningGoalContext' }
        constraints: { $ref: '#/components/schemas/Constraints' }
        requestedCapability:
          type: string
          maxLength: 64
...
    AIProposalEnvelope:
      type: object
      additionalProperties: false
      required:
        [contractVersion, proposalId, agentType, agentVersion, modelRoute, trustLevel, proposal]
      properties:
        contractVersion: { $ref: '#/components/schemas/ContractVersion' }
        proposalId:
          type: string
          maxLength: 64
          description: Identity of this proposal. Referenced by DecisionRecord when it influences one.
        agentType: { $ref: '#/components/schemas/AgentType' }
        agentVersion: { type: string, maxLength: 64 }
        agentRunId:
          type: string
          maxLength: 64
          description: >
            The orchestrated agent execution that produced this proposal (Observability HLD §9).
            Distinct from proposalId, requestId and interactionId: a retried request produces
            several runs, and one interaction may involve several agents. Carried on the wire so
            the deterministic core's log lines for a decision can name the run that proposed it --
            otherwise the correlation chain ends at the plane boundary, which is where a support
            pivot most often needs to cross. Optional, so a plane that predates the field still
            validates.
        promptTemplateId:
          type: string
          maxLength: 64
          description: >
            Which prompt produced this proposal (M1-ADR-011). Two of the assessment agent's prompts
            share a version, so promptVersion alone does not identify one.
        promptVersion: { type: string, maxLength: 64 }
        modelRoute: { type: string, maxLength: 64 }
        resolvedProvider:
          type: string
          maxLength: 64
          description: Provider resolved by the governed route table; never supplied by the caller.
        modelId:
          type: string
          maxLength: 128
          description: Immutable concrete model identifier used for this execution.
        routeVersion:
          type: string
          maxLength: 256
          description: Version/configuration stamp of the route resolution used for this execution.
        trustLevel: { $ref: '#/components/schemas/TrustLevel' }
        confidence:
          $ref: '#/components/schemas/DecimalString'
          description: >
            The agent's self-reported confidence. Never a basis for authority: Spring's deterministic
            policy decides, and a high value here changes nothing.
        reasonCodes:
          type: array
          maxItems: 16
          items: { type: string, maxLength: 64 }
        proposal:
          type: object
          description: >
            Agent-specific payload. Deliberately open at v1.0 so each agent can define its own shape
            in its own task without a breaking contract change; each agent's schema is pinned by its
            own golden fixtures.
          additionalProperties: true
        validation: { $ref: '#/components/schemas/Validation' }
        usage: { $ref: '#/components/schemas/Usage' }
...
    DiagnosticAssessmentRequest:
      type: object
      additionalProperties: false
      required: [contractVersion, interactionId, requestId, constraints, groundedContext]
      description: >
        Carries correlation identity and the grounded context, and nothing else. There is no learner
        reference beyond the opaque one inside the context, and no capability string: this operation
        does one thing.
      properties:
        contractVersion: { $ref: '#/components/schemas/ContractVersion' }
        interactionId:
          type: string
          maxLength: 64
          description: Logical learner action. Stable across safe retries.
        requestId:
          type: string
          maxLength: 64
          description: One transport attempt. New per retry.
        constraints: { $ref: '#/components/schemas/Constraints' }
        groundedContext: { $ref: '#/components/schemas/GroundedContextEnvelope' }
...
    AssessmentEvaluationRequest:
      type: object
      additionalProperties: false
      required: [contractVersion, interactionId, requestId, constraints, evaluationContext,
                 groundedContext]
      description: >
        M2-T11 request with no natural learner identifier and no authority-bearing action. The
        typed evaluation context is bound to the authorized grounded context by evidence IDs.
      properties:
        contractVersion: { $ref: '#/components/schemas/ContractVersion' }
        interactionId:
          type: string
          minLength: 1
          maxLength: 64
          description: Logical learner action. Stable across safe retries.
        requestId:
          type: string
          minLength: 1
          maxLength: 64
          description: Stable logical idempotency identity for this evaluation.
        constraints: { $ref: '#/components/schemas/Constraints' }
        evaluationContext: { $ref: '#/components/schemas/AssessmentEvaluationContext' }
        groundedContext: { $ref: '#/components/schemas/GroundedContextEnvelope' }
...